template_bot/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-debug-toolbar==1.6
djangorestframework==3.5.4
future==0.16.0
gunicorn==19.6.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
sqlparse==0.2.2
